--- a/.github/workflows/test.yaml
+++ b/.github/workflows/test.yaml
@@ -12,14 +12,12 @@ jobs:
     name: Linux
     runs-on: ubuntu-latest
     steps:
-    - name: Set up Go 1.14
+    - name: Set up Go 1.19
       uses: actions/setup-go@v2
       with:
-        go-version: '1.14.2'
+        go-version: '1.19'
       id: go
     - name: Check out code into the Go module directory
       uses: actions/checkout@v2
-    - name: Install golint
-      run: go get -u golang.org/x/lint/golint
     - name: Test
       run: "./test"