Fixes #346

```
% benchstat ~/old.txt ~/new.txt 
name      old time/op  new time/op  delta
Verify-8  35.5µs ± 3%  32.6µs ± 2%  -8.30%  (p=0.000 n=10+10)
```